Doctors Management + Mail

+++ resources/views/_health_institution/message_listing.blade.php
@@ -8,16 +8,18 @@
     <div class="row">
         <div class="col-12">
             <div class="card">
-                <div class="card-header">
-                    <a class="heading-elements-toggle">
-                        <i class="la la-ellipsis-v font-medium-3"></i>
-                    </a>
-                    <div class="heading-elements">
-                        <a href="{{ route('institution_messages.create') }}" class="btn btn-info btn-sm">
-                            <i class="ft-plus white"></i> Add Message
+                @can('isCountryHead')
+                    <div class="card-header">
+                        <a class="heading-elements-toggle">
+                            <i class="la la-ellipsis-v font-medium-3"></i>
                         </a>
+                        <div class="heading-elements">
+                            <a href="{{ route('institution_messages.create') }}" class="btn btn-info btn-sm">
+                                <i class="ft-plus white"></i> Add Message
+                            </a>
+                        </div>
                     </div>
-                </div>
+                @endcan
                 <div class="card-content">
                     <div class="card-body card-dashboard">
                         <table class="table table-striped table-bordered dtTable">
@@ -36,18 +38,23 @@
                                         <td> {{ $message->title }}</td>
                                         <td> {{ $message->content }} </td>
                                         <td class="no-wrap">
-                                            <a href="{{ route('institution_messages.edit', $message->uuid ) }}" class="btn btn-icon btn-info btn-sm"> <i class="la la-edit"></i> </a>
 
-                                            {!! Form::open(array(
-                                                    'route' => array('institution_messages.delete', $message->uuid),
-                                                    'method' => 'delete',
-                                                    'class'=>'delSwalForm',
-                                                    'style'=>'display:inline'
-                                            )) !!}
+                                            <a href="{{ route('institution_messages.triggerPushMessage', $message->uuid ) }}" class="btn btn-icon btn-success btn-sm"> <i class="la la-paper-plane"></i> Send </a>
 
-                                            <button type="submit" class="btn btn-icon btn-danger btn-sm delSwal"> <i class="la la-trash"></i> </button>
+                                            @can('isCountryHead')
+                                                <a href="{{ route('institution_messages.edit', $message->uuid ) }}" class="btn btn-icon btn-info btn-sm"> <i class="la la-edit"></i> </a>
 
-                                            {!! Form::close() !!}
+                                                {!! Form::open(array(
+                                                        'route' => array('institution_messages.delete', $message->uuid),
+                                                        'method' => 'delete',
+                                                        'class'=>'delSwalForm',
+                                                        'style'=>'display:inline'
+                                                )) !!}
+
+                                                <button type="submit" class="btn btn-icon btn-danger btn-sm delSwal"> <i class="la la-trash"></i> </button>
+
+                                                {!! Form::close() !!}
+                                            @endcan
                                         </td>
                                     </tr>
                                 @endforeach
